+++ changelog.md
@@ -2,10 +2,15 @@
 
 Non-tradeable items that's ignored or blacklisted are no longer recorded in the loot history (#210).
 
+### Dev
+
+- *Added itemLink to `RCMLAwardSuccess` and `RCMLAwardFailed` AceEvents.*
+
 ## Bugfixes
 
 - *Fixed issue with Auto Award BoE list. (curse#500)*
 - *Fixed occasional missing data error. (curse#503)*
+
 
 # v3.3.0